+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/storage/portablegrid/PortableGridScreen.java
@@ -1,5 +1,6 @@
 package com.refinedmods.refinedstorage.common.storage.portablegrid;
 
+import com.refinedmods.refinedstorage.common.Platform;
 import com.refinedmods.refinedstorage.common.grid.screen.AbstractGridScreen;
 import com.refinedmods.refinedstorage.common.support.widget.ProgressWidget;
 
@@ -18,9 +19,11 @@ public class PortableGridScreen extends AbstractGridScreen<AbstractPortableGridC
     private static final int DISK_SLOT_HEIGHT = 26;
 
     private static final ResourceLocation TEXTURE = createIdentifier("textures/gui/portable_grid.png");
+    private static final ResourceLocation NO_ENERGY_TEXTURE = createIdentifier("textures/gui/grid.png");
 
     @Nullable
     private ProgressWidget progressWidget;
+    private final ResourceLocation texture;
 
     public PortableGridScreen(final AbstractPortableGridContainerMenu menu,
                               final Inventory inventory,
@@ -29,6 +32,9 @@ public class PortableGridScreen extends AbstractGridScreen<AbstractPortableGridC
         this.inventoryLabelY = 75;
         this.imageWidth = 193;
         this.imageHeight = 176;
+        this.texture = Platform.INSTANCE.getConfig().isRequireEnergy()
+            ? TEXTURE
+            : NO_ENERGY_TEXTURE;
     }
 
     @Override
@@ -58,6 +64,9 @@ public class PortableGridScreen extends AbstractGridScreen<AbstractPortableGridC
             DISK_SLOT_WIDTH,
             DISK_SLOT_HEIGHT
         ));
+        if (!Platform.INSTANCE.getConfig().isRequireEnergy()) {
+            return;
+        }
         final int progressX = 172;
         final int progressY = imageHeight - 10 - 70;
         if (progressWidget == null) {
@@ -81,11 +90,11 @@ public class PortableGridScreen extends AbstractGridScreen<AbstractPortableGridC
         super.renderBg(graphics, delta, mouseX, mouseY);
         final int x = (width - imageWidth) / 2;
         final int y = (height - imageHeight) / 2;
-        graphics.blit(getTexture(), x - DISK_SLOT_WIDTH + 3, y + 3, 226, 0, DISK_SLOT_WIDTH, DISK_SLOT_HEIGHT);
+        graphics.blit(TEXTURE, x - DISK_SLOT_WIDTH + 3, y + 3, 226, 0, DISK_SLOT_WIDTH, DISK_SLOT_HEIGHT);
     }
 
     @Override
     protected ResourceLocation getTexture() {
-        return TEXTURE;
+        return texture;
     }
 }